first version of admin route
